requirements.txt
tensorflow
numpy
matplotlib
pandas
scikit-learn
opencv-python
albumentations
pillow